--- frontend/src/screens/dashboard/docsqa/main/partials/index.tsx
@@ -1,15 +1,0 @@
-import ConfigSelector from './ConfigSelector'
-import ApplicationModal from './ApplicationModal'
-import PromptForm from './PromptForm'
-import NoAnswer from './NoAnswer'
-import ErrorAnswer from './ErrorAnswer'
-import Answer from './Answer'
-
-export {
-  ConfigSelector,
-  ApplicationModal,
-  PromptForm,
-  NoAnswer,
-  ErrorAnswer,
-  Answer,
-}
